slackへメッセージを飛ばす機能を関数にして、別ファイルで管理するようにしました。
